fix:Correcting index.ts and dividing 'converTo' and 'download' functions into different TS files
